Extract SitemapExtension into seperate file
